pyproject.toml
[project]
name = "72admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "captcha>=0.7.1",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-currentuser>=0.9.0",
    "django-ninja>=1.4.5",
    "django-ninja-extra>=0.30.2",
    "django-ninja-jwt>=5.4.0",
    "django-redis>=6.0.0",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.2.12",
    "pyjwt>=2.10.1",
    "redis>=7.0.1",
]

[tool.basedpyright]
typeCheckingMode = "off"
